Make current_partial_gop int64_t
This makes it easier to track wraparounds etc. Note that the
value is still stored as an uint32_t in the TLV.

+++ lib/src/sv_auth.c
@@ -89,7 +89,7 @@ decode_sei_data(signed_video_t *self, const uint8_t *payload, size_t payload_siz
 {
   assert(self && payload && (payload_size > 0));
   gop_info_t *gop_info = self->gop_info;
-  int64_t partial_gop_number = (int64_t)gop_info->current_partial_gop;
+  int64_t partial_gop_number = gop_info->current_partial_gop;
   DEBUG_LOG("SEI payload size = %zu, exp (partial) gop number = %ld", payload_size,
       gop_info->latest_validated_gop + 1);
 
@@ -100,7 +100,7 @@ decode_sei_data(signed_video_t *self, const uint8_t *payload, size_t payload_siz
   }
 
   // Compare new with last number of GOPs to detect potential wraparound.
-  int64_t new_partial_gop_number = (int64_t)gop_info->current_partial_gop;
+  int64_t new_partial_gop_number = gop_info->current_partial_gop;
   if (new_partial_gop_number < partial_gop_number) {
     // There is a potential wraparound, but it could also be due to re-ordering of SEIs.
     // Use the distance to determine which of these options is the most likely one.
@@ -122,7 +122,7 @@ detect_lost_sei(signed_video_t *self)
   // Get the last GOP counter.
   int64_t exp_partial_gop_number = gop_info->latest_validated_gop + 1;
   // Compare new with last number of GOPs to detect potentially lost SEIs.
-  int64_t new_partial_gop_number = (int64_t)gop_info->current_partial_gop;
+  int64_t new_partial_gop_number = gop_info->current_partial_gop;
   // Compensate for counter wraparounds.
   new_partial_gop_number += (int64_t)gop_info->num_partial_gop_wraparounds << 32;
   int64_t potentially_lost_seis = new_partial_gop_number - exp_partial_gop_number;
